requirements.txt
aiogram == 3.*
aiohttp
python-dotenv
pydantic-settings
requests == 2.32.5
pandas>=2.2,<2.4
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
numpy==1.25.*
